% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_world.R
\name{abc}
\alias{abc}
\title{Alphabet}
\usage{
abc(case)
}
\arguments{
\item{case}{lower ("lc") or upper case ("uc")}
}
\value{
the alphabet string
}
\description{
Return the alphabet
}
\examples{
abc("uc");
}
