% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{actual_vs_predicted}
\alias{actual_vs_predicted}
\title{Actual vs. Predicted Plot and Residual Plot}
\usage{
actual_vs_predicted(trainer, predictions)
}
\arguments{
\item{trainer}{The trainer object containing the test data and actual labels}

\item{predictions}{The predicted values obtained from the trained model}
}
\value{
A list containing the following plots:
\item{actual_predicted}{Actual vs. Predicted plot}
\item{residuals_plot}{Residual plot}
}
\description{
This function generates two plots to compare the actual values with the predicted values and visualize the residuals.
The first plot shows the actual values on the y-axis and the predicted values on the x-axis, with points indicating the data points.
The second plot displays the residuals (the differences between the actual and predicted values) on the y-axis and the predicted values on the x-axis.
The residual plot helps analyze the distribution of residuals and identify any patterns or trends.
}
