% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add.R
\name{Add}
\alias{Add}
\title{Add two numbers}
\usage{
Add(x, y)
}
\arguments{
\item{x}{is a numeric value}

\item{y}{is a numeric value}
}
\value{
The sum of the 2 inputs
}
\description{
Takes in two numbers to add up
}
