% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_two.R
\name{add_two}
\alias{add_two}
\title{add_two}
\usage{
add_two(a, b)
}
\arguments{
\item{a}{An integer}

\item{b}{Another integer}
}
\value{
the sum of the two integers
}
\description{
add_two
}
\examples{
sum(1, 2)
3
}
