% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{apply_log_transform}
\alias{apply_log_transform}
\title{Apply Logarithm Transformation to a Specific Column}
\usage{
apply_log_transform(data, column)
}
\arguments{
\item{data}{The input data frame.}

\item{column}{The name of the column to apply the transformation on.}
}
\value{
The data frame with the logarithm transformation applied to the specified column (if it is numeric).
}
\description{
This function applies the natural logarithm transformation to a specific column in a data frame, if the column is numeric.
}
\examples{
df <- data.frame(a = c(1, 2, 3), b = c(4, 5, 6), c = c("x", "y", "z"))
df_transformed <- apply_log_transform(df, "a")

}
