% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_of_circle.R
\name{area_of_circle}
\alias{area_of_circle}
\title{Circle area}
\usage{
area_of_circle(radius)
}
\arguments{
\item{radius}{the radius of the circle}
}
\description{
Function which calculates the area of a circle
}
\author{
Eveline van der Linden
}
