% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_of_square.R
\name{area_of_square}
\alias{area_of_square}
\title{Square area}
\usage{
area_of_square(length, width)
}
\arguments{
\item{length}{the length of the square}

\item{width}{the width of the square}
}
\description{
Function which calculates the area of a square
}
\author{
Eveline van der Linden
}
