% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AreaOfSquares.R
\name{AreaOfSquares}
\alias{AreaOfSquares}
\title{Square area}
\usage{
AreaOfSquares(length, width)
}
\arguments{
\item{length}{The length of the square}

\item{width}{The widht of the square}
}
\description{
This function calculates the area of a square.
}
\author{
Daniel van de Craats
}
