% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BVAR_Jeff.R
\name{BVAR_Jeff}
\alias{BVAR_Jeff}
\title{Draws posterior of VAR(p) Model with non-informative prior}
\usage{
BVAR_Jeff(Yts, p, const)
}
\arguments{
\item{Yts}{n x T matrix of data.}

\item{p}{the lag lenght of the VAR.}

\item{const}{A scalar, where const = 1 for model with intercept, const = 0 for model without intercept.}
}
\value{
A List with:
\itemize{
\item betar_all n x np+1 x R-burn matrices of coefficients.
\item Sigr_all n x n x R-burn matrices of variance covariance matrices.
}
}
\description{
\code{create_inventory} constructs companion matrix for VAR(p) model.
}
\details{
This function draws from the posterior of a VAR(p) model with non-informatibve prior
}
\examples{
param <- BVAR_Jeff(ft,q,0)
}
