% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calc_balanced_metrics_classification}
\alias{calc_balanced_metrics_classification}
\title{Calculate Balanced Metrics for Classification}
\usage{
calc_balanced_metrics_classification(labels, predicted)
}
\arguments{
\item{labels}{The true labels}

\item{predicted}{The predicted labels}
}
\value{
A list containing the calculated metrics:
\item{precision_pos}{Precision for the positive class}
\item{recall_pos}{Recall for the positive class}
\item{f1_pos}{F1-score for the positive class}
\item{precision_neg}{Precision for the negative class}
\item{recall_neg}{Recall for the negative class}
\item{f1_neg}{F1-score for the negative class}
\item{balanced_precision}{Balanced precision}
\item{balanced_recall}{Balanced recall}
\item{balanced_f1}{Balanced F1-score}
}
\description{
This function calculates various balanced metrics for a binary classification problem,
including precision, recall, and F1-score for both positive and negative classes,
as well as balanced precision, balanced recall, and balanced F1-score.
}
