% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calc_metrics_regression}
\alias{calc_metrics_regression}
\title{Calculate Metrics for Regression}
\usage{
calc_metrics_regression(trainer, model)
}
\arguments{
\item{trainer}{The regression trainer object}

\item{model}{The trained regression model}
}
\value{
A list containing the calculated metrics:
\item{predicted_values}{Predicted values for the test data}
\item{train_mse}{Training mean squared error}
\item{train_rmse}{Training root mean squared error}
\item{train_mae}{Training mean absolute error}
\item{train_rsquared}{Training coefficient of determination (R-squared)}
\item{test_mse}{Test mean squared error}
\item{test_rmse}{Test root mean squared error}
\item{test_mae}{Test mean absolute error}
\item{test_rsquared}{Test coefficient of determination (R-squared)}
}
\description{
This function calculates various evaluation metrics for a regression problem,
including mean squared error (MSE), root mean squared error (RMSE),
mean absolute error (MAE), and coefficient of determination (R-squared).
}
