% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcTSNE.R, R/DimensionReduction.R
\name{CalcTSNE}
\alias{CalcTSNE}
\title{Returns the Y cordinates for Tsne plotting.}
\usage{
CalcTSNE(sco, ndim)

CalcTSNE(sco, ndim)
}
\arguments{
\item{sco}{single cell object S4 file}

\item{sco}{single cell object S4 file}
}
\description{
Returns the Y cordinates for Tsne plotting.

Returns the Y cordinates for Tsne plotting.
}
