% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_learning_curve}
\alias{calculate_learning_curve}
\title{Calculate Learning Curve}
\usage{
calculate_learning_curve(trainer)
}
\arguments{
\item{trainer}{The trainer object containing the training and test data}
}
\value{
A list of metrics at different training data sizes:
If the status is "classification":
\item{train_acc}{Training accuracy}
\item{train_prec}{Training precision}
\item{train_rec}{Training recall}
\item{train_f1}{Training F1-score}
\item{test_acc}{Test accuracy}
\item{test_prec}{Test precision}
\item{test_rec}{Test recall}
\item{test_f1}{Test F1-score}

If the status is "regression":
\item{train_mse}{Training mean squared error}
\item{train_rmse}{Training root mean squared error}
\item{train_mae}{Training mean absolute error}
\item{train_rsquared}{Training coefficient of determination (R-squared)}
\item{test_mse}{Test mean squared error}
\item{test_rmse}{Test root mean squared error}
\item{test_mae}{Test mean absolute error}
\item{test_rsquared}{Test coefficient of determination (R-squared)}
}
\description{
This function calculates the learning curve for a classification or regression model.
It iteratively trains the model on different subsets of the training data with increasing sizes,
and evaluates the model's performance on both the training and test data at each iteration.
The calculated metrics differ based on the status of the trainer object.
For classification models, the metrics include accuracy, precision, recall, and F1-score.
For regression models, the metrics include mean squared error (MSE), root mean squared error (RMSE),
mean absolute error (MAE), and coefficient of determination (R-squared).
}
