% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_numeric_column_means}
\alias{calculate_numeric_column_means}
\title{Calculate Mean for Numeric Columns}
\usage{
calculate_numeric_column_means(data)
}
\arguments{
\item{data}{The input data frame.}
}
\value{
A vector containing the mean values for each numeric column.
}
\description{
This function calculates the mean for each numeric column in a data frame.
}
\examples{
df <- data.frame(a = c(1, 2, 3), b = c(4, 5, 6))
means <- calculate_numeric_column_means(df)

}
