% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{chan.pval}
\alias{chan.pval}
\title{Chan's Exact P-value}
\usage{
chan.pval(x.T, x.C, N.T, N.C, Delta0, method, lower = TRUE, tol = 0.001, ...)
}
\arguments{
\item{x.T}{observed number of responders in Treatment group}

\item{x.C}{observed number of responders in Control group}

\item{N.T}{sample size in Treatment group}

\item{N.C}{sample size in Control group}

\item{Delta0}{non-inferiority margin}

\item{method}{method for ordering criterion}

\item{lower}{TRUE for the null hypothesis P.T-P.C<=-Delta0, FALSE for the null hypothesis P.T-P.C>-Delta0}

\item{tol}{increment size for domain of Delta, default set to 0.001}
}
\value{
Exact p-value
}
\description{
This function calculates the exact p-value based on Chan (1998)
}
