% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{chan.zhang}
\alias{chan.zhang}
\title{Chan and Zhang Confidence Interval}
\usage{
chan.zhang(
  x.T,
  x.C,
  N.T,
  N.C,
  Delta0,
  method,
  alpha = 0.05,
  tol = 0.001,
  width = 0.3
)
}
\arguments{
\item{x.T}{observed number of responders in Treatment group}

\item{x.C}{observed number of responders in Control group}

\item{N.T}{sample size in Treatment group}

\item{N.C}{sample size in Control group}

\item{Delta0}{non-inferiority margin}

\item{method}{method for ordering criterion}

\item{alpha}{significance level}

\item{tol}{increment size for domain of Delta, default set to 0.001}

\item{width}{range from starting values based on Miettinen & Nurminen confidence limits}
}
\value{
Chan and Zhang Confidence Interval
}
\description{
This function calculates the confidence interval based on Chan and Zhang (1999)
}
