% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Circumference.R
\name{circumference}
\alias{circumference}
\title{Circle Circumference}
\usage{
circumference(r)
}
\arguments{
\item{r}{Numeric, radius measurement of a circle}
}
\value{
circumference
}
\description{
This function calculates the circumference of a circle
}
\examples{
circumference(10)
}
