% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{Cluster}
\alias{Cluster}
\title{Clusters and split the data into N partitions}
\usage{
Cluster(sco, N)
}
\arguments{
\item{sco}{scells object}

\item{N}{N number of groups}
}
\description{
Clusters and split the data into N partitions
}
