% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{clusterData}
\alias{clusterData}
\title{Clusters the data and splits the dendrogram into k clusters}
\usage{
clusterData(sco, k)
}
\arguments{
\item{sco}{An scell S4 object}

\item{k}{The number of clusters we wish to retrieve}
}
\description{
Clusters the data and splits the dendrogram into k clusters
}
