% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphs.R
\name{ColourTSNEbyClust}
\alias{ColourTSNEbyClust}
\title{Makes a 3D tSNE plot where cells are coloured by cluster}
\usage{
ColourTSNEbyClust(sco)
}
\description{
Makes a 3D tSNE plot where cells are coloured by cluster
}
