% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphs.R
\name{ColourTSNEbyGeneExpr}
\alias{ColourTSNEbyGeneExpr}
\title{Makes a 3D tSNE plot where cells are coloured by the expression of a given gene}
\usage{
ColourTSNEbyGeneExpr(sco, gene)
}
\description{
Makes a 3D tSNE plot where cells are coloured by the expression of a given gene
}
