% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compFstate.R
\name{compFstate}
\alias{compFstate}
\title{Construct companion matrix for VAR(p) model}
\usage{
compFstate(phi, Q, H, con)
}
\arguments{
\item{phi}{Diagoanl matrix of dimension k x k with vector autoregressive coefficients.}

\item{Q}{A matrix of .}

\item{H}{A matrix of lambadas from Multi-Move Gibbs sampler in companion form.}

\item{con}{A scalar, where con = 1 for model with intercept, con = 0 for model without intercept.}
}
\value{
A List with phi, Q, H in companion form.
}
\description{
\code{create_inventory} constructs companion matrix for VAR(p) model.
}
\details{
The function runs inside the Multi-Move Gibbs sampler.
}
\examples{
matcomp <- compFstate(phi,Q,lambda,const)
}
