% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{confint.z}
\alias{confint.z}
\title{Chan's Exact P-value}
\usage{
\method{confint}{z}(x.T, x.C, N.T, N.C, Delta0, method, EC, alpha = 0.05, tol = 1e-10, ...)
}
\arguments{
\item{x.T}{observed number of responders in Treatment group}

\item{x.C}{observed number of responders in Control group}

\item{N.T}{sample size in Treatment group}

\item{N.C}{sample size in Control group}

\item{Delta0}{non-inferiority margin}

\item{method}{method for ordering criterion}

\item{EC}{TRUE if confidence limits are exact-corrected to match Chan's exact p-value}

\item{tol}{tolerance for convergence}
}
\value{
List of lower and upper confidence limits (D.lower, D.upper)
}
\description{
This function computes the confidence interval limits given an ordering criterion with or
without the exact-correction
}
