% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions.R
\name{convertKNMI.fun}
\alias{convertKNMI.fun}
\title{Convert KNMI input}
\usage{
convertKNMI.fun(Datafile)
}
\arguments{
\item{Datafile}{Name of KNMI data file to read}
}
\description{
This function converts standard KNMI input datafiles to readable input files for the main SUTRA script.
   Function not to be called outside of main program.
}
