% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_random_forest}
\alias{create_random_forest}
\title{Create a random forest model}
\usage{
create_random_forest(trainer, formula, num_trees, mtry, min_node_size)
}
\arguments{
\item{trainer}{An instance of the classification or regression trainer}

\item{formula}{The formula specifying the model formula}

\item{num_trees}{The number of trees to grow}

\item{mtry}{The number of variables randomly sampled as candidates at each split}

\item{min_node_size}{The minimum number of observations in a terminal node}
}
\value{
A list containing the trained model and additional information
}
\description{
This function creates a random forest model using the ranger algorithm for
classification or regression tasks.
}
