% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_trainer_regression}
\alias{create_trainer_regression}
\title{Create an instance of the regression trainer}
\usage{
create_trainer_regression(train_data, test_data, label_column)
}
\arguments{
\item{train_data}{Training data that the user wants to use for training}

\item{test_data}{Testing data that the user wants to use for testing}

\item{label_column}{String with a name of the dependent variable}
}
\value{
An instance of the regression trainer
}
\description{
Create an instance of the regression trainer
}
