% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{cube}
\alias{cube}
\title{Cube a number}
\usage{
cube(x)
}
\arguments{
\item{x}{A numeric value to be cubed}
}
\value{
the cube of the input
}
\description{
Takes in any numeric value and cube it
}
\examples{
cube(1)
cube(10)
}
