 \name{dataset}
\alias{dataset}
\docType{data}
\title{
Vector normally distributed
}
\description{
Random vector generated following a normal distribution. Length equal to 1,000.
}
\usage{data("dataset")}
\format{
  The format is:
 num [1:1000] 0.811 0.241 2.508 0.956 0.486 ...
}

\examples{
data(dataset)
}
\keyword{datasets}
