% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df}
\alias{df}
\title{Lengths and widths
This dataset contains uniformaly distributed random values of lengths and widths}
\format{15 rows and columns
\describe{
\item{length}{length of the rectangle}
\item{width}{width of the rectangle}
}}
\usage{
df
}
\description{
Lengths and widths
This dataset contains uniformaly distributed random values of lengths and widths
}
\keyword{datasets}
