% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation.R
\name{difference}
\alias{difference}
\title{Difference of two numeric values}
\usage{
difference(x, y)
}
\arguments{
\item{x}{A numeric value}

\item{y}{A numeric value}
}
\value{
the difference of these two numeric values
}
\description{
Difference of two numeric values
}
\examples{
difference(2,1)
difference(1,2)
}
