% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mat.R
\name{divide}
\alias{divide}
\title{Divide Math}
\usage{
divide(a, b)
}
\arguments{
\item{a}{math parameter 1}

\item{b}{math parameter 2}
}
\value{
result
}
\description{
This function returns a result divide two parameters
}
