% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublemean.R
\name{doublemean}
\alias{doublemean}
\title{My doublemean function.}
\usage{
doublemean(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric or a string "It's significant"
}
\description{
My doublemean returns a numeric or a string "It's significant".
}
\examples{
doublemean(1:6)
}
