% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advRlab3.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A number}

\item{b}{A number}
}
\value{
The greatest common denominator of \code{a} and \code{b}
}
\description{
Euclidean algorithm
}
\examples{
euclidean(10,100)
euclidean(25,125)
\\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
