% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/even_or_odd.R
\name{even_or_odd}
\alias{even_or_odd}
\title{Test Whether a Number is Even or Odd}
\usage{
even_or_odd(number)
}
\description{
This function tests whether a number is even or odd.
When the remainder of the number divided by 2 is 0, it is an even number.
When the remainder of the number divided by 2 is 1, it is an odd number.
When the remainder of the number divided by 2 is something else,
the number is not an integer.
If the input is not of numeric type, the function returns the message that
it is not a number.
}
\examples{
even_or_odd(20)
even_or_odd(25)
even_or_odd(13.5)
even_or_odd("square")
}
\keyword{number}
