% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Week 2 assignment.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Mapping the accident for any given state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{state code}

\item{year}{accident year}
}
\value{
a map showing the location of the accidents
}
\description{
Mapping the location of the accidents (based on latitude and longitude) with user speicifed state number
and accident year
}
\note{
the function will require "maps" package
}
\examples{
fars_map_state(10,2013)
}
