% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Week 2 assignment.R
\name{fars_read}
\alias{fars_read}
\title{Read Data from Fatality Analysis Reporting System}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{the path to the data}
}
\value{
the wrapped dataset
}
\description{
Read data from a csv file. The function also uses the tbl_df wrapper from dplyr to make sure
the system won't print a lot of data to the screen
}
\note{
The data has to be in csv format
}
