% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Week 2 assignment.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read the year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector of your desired accident years}
}
\value{
a tibble with month and year
}
\description{
Read user specified year for the FARS data
}
\examples{
\dontrun{fars_read_years(2013)}
}
