% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Week 2 assignment.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize the FARS data with specified accident year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{year}{accident year}
}
\value{
a tibble showing number of accident by month
}
\description{
Summarize number of accidents across each month for the specified year
}
\examples{
\dontrun{fars_summarize_years(2013)}
}
