% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_get_conc.R
\name{get_conc}
\alias{get_conc}
\title{get a new concentration (title)}
\usage{
get_conc(conc, OD, new, ...)
}
\arguments{
\item{conc}{takes a numeric vector with concentrations}

\item{OD}{takes optical densities}

\item{new}{takes vector with...}

\item{...}{currently disregarded}
}
\value{
the predictions
}
\description{
It calculates a new concentration based on a calbration curve (description).
 It uses internally the function \code{\link{calib}}.
}
\details{
here you can give more details

those details they can be multiple paragraphs
}
\examples{
conc <- 1:10
OD <- 10:1
get_conc(conc, OD, 4.5)

}
\seealso{
\code{\link{calib}}.
}
