% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrds.R
\name{getPrds}
\alias{getPrds}
\title{getPrds}
\usage{
getPrds(start_date, end_date = Sys.Date() - days(1))
}
\arguments{
\item{end_date}{}
}
\value{

}
\description{
This function returns calendar month periods for `map` GA queries
}
