% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetVarGenes.R
\name{GetVarGenes}
\alias{GetVarGenes}
\title{Calculates the top N variable genes from a scell object}
\usage{
GetVarGenes(sco, nvar)
}
\arguments{
\item{sco}{An scell S4 object}

\item{nvar}{The number of genes we wish to retrieve}
}
\description{
Calculates the top N variable genes from a scell object
}
