% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{\%>\%}
\alias{\%>\%}
\alias{my_rf_cv}
\title{Random Forest function}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{The number of folds desired}
}
\value{
A numeric with the cross-validation error
}
\description{
This function performs predicts \code{body_mass_g} using covariates
\code{bill_length_mm}, \code{bill_depth_mm}, and \code{flipper_length_mm}
using random forests.
}
\examples{
my_rf_cv(5)
my_rf_cv(20)

}
\keyword{prediction}
