% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/happy_function.R
\name{happy_function}
\alias{happy_function}
\title{A Happy Function}
\usage{
happy_function(happy = TRUE)
}
\arguments{
\item{happy}{Are you happy? Defaults to TRUE.}
}
\description{
This function allows you to show if you are happy or angry.
}
\examples{
happy_function()
}
\keyword{happy}

