% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatMap.R
\name{HeatMapVarGenes}
\alias{HeatMapVarGenes}
\title{Generate a heatmap of N variable genes determined in GetVarGenes}
\usage{
HeatMapVarGenes(sco)
}
\arguments{
\item{sco}{An scell S4 object}
}
\description{
Generate a heatmap of N variable genes determined in GetVarGenes
}
