% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{hello_world}
\alias{hello_world}
\title{Hello World to 'myfirstpackage' Users}
\usage{
hello_world(name = "pen")
}
\arguments{
\item{name}{character, will replace \code{name} in the displayed text}
}
\value{
Returns the printed message
}
\description{
Displays "Hello world, I have one name", where \code{name}
can be replaced.
}
\details{
The details are highly encouraged, but not mandatory
}
\examples{

hello_world("apple")
hello_world("pen")
hello_world("penpal apple apple pen")

}
