% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{howay}
\alias{howay}
\title{Function that uses div}
\usage{
howay(x, y, z)
}
\arguments{
\item{x}{first number}

\item{y}{second number}

\item{z}{third number}
}
\description{
divide then times
}
\examples{
howay(1,2,4)
howay(3,2,1)
}
