% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KK_QAC_May2024_samples.R
\docType{data}
\name{KK_QAC_May2024_samples}
\alias{KK_QAC_May2024_samples}
\title{KK_QAC_May2024_samples Dataset}
\format{
A data frame with 116 rows and 8 columns:
\describe{
\item{Sample ID}{ This is a unique number for each sample }
\item{Study ID}{ this number shows that we are working on which project and study }
\item{date sample was taken YYYY-MM-DD}{ the exact date that we took our samples}
\item{Study Name}{ this is the project name }
\item{Sample Type}{ this will show which animal or organism do we work on it }
\item{Date Received YYYY-MM-DD}{ the exact date that we recieved our samples }
\item{Notes}{ any note which might be important }
\item{Illumina Plate Number}{ an specific number }
...
}
}
\source{
Mesocosm .
}
\usage{
KK_QAC_May2024_samples
}
\description{
This dataset contains samples about microbiomes of Mesocosm experiment on 2023 for QAC project .
}
\keyword{datasets}
