% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadData.R
\name{LoadData}
\alias{LoadData}
\title{reads the input file and makes an scell object immediately}
\usage{
LoadData(InputFile)
}
\arguments{
\item{InputFile}{any input file containing expression data}
}
\description{
reads the input file and makes an scell object immediately
}
