% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Week 2 assignment.R
\name{make_filename}
\alias{make_filename}
\title{Make a file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{your desired accident year}
}
\value{
a character string of the name of the file
}
\description{
Make a file name with speicified accident year
}
\examples{
make_filename(2013)
}
