% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makecluster.R
\name{makecluster}
\alias{makecluster}
\title{Make heatmap of top variable genes and return clusters}
\usage{
makecluster(sco, N)
}
\arguments{
\item{sco}{input variable gene matrix}
}
\description{
Make heatmap of top variable genes and return clusters
}
