% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeheatmap.R
\name{makeheatmap}
\alias{makeheatmap}
\title{Make heatmap of top variable genes}
\usage{
makeheatmap(sco)
}
\arguments{
\item{sco}{input variable gene matrix}
}
\description{
Make heatmap of top variable genes
}
