% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makescellclass.R
\name{makescellclass}
\alias{makescellclass}
\title{Read delimeited text file into sclass-object}
\usage{
makescellclass(myfilename)
}
\arguments{
\item{myfilename}{input filename}
}
\description{
Read delimeited text file into sclass-object
}
