% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplyr.R
\name{multiplyr}
\alias{multiplyr}
\title{A function for multiplying two numbers}
\usage{
multiplyr(a = 1, b = 5)
}
\arguments{
\item{a}{A real number}

\item{b}{A real number}
}
\value{
The product of \code{a} and \code{b}
}
\description{
A baby function to illustrate how to document stuff in R
It is simply a function to multiply two numbers
}
\examples{
multiplyr(2, 5)
multiplyr(4, 8)
}
