% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{my.clusters}
\alias{my.clusters}
\title{Returns the Y cordinates for Tsne plotting.}
\usage{
my.clusters(sco, nclust)
}
\arguments{
\item{sco}{single cell object S4 file}

\item{nclust}{Number of clusters}
}
\description{
Returns the Y cordinates for Tsne plotting.
}
