% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{my.df}
\alias{my.df}
\title{Lengths and widths}
\format{A data frame with 15 rows and 2 variables:
\describe{
\item{length}{length of the square}
\item{width}{width of the square}
}}
\usage{
my.df
}
\description{
This dataset contains random numerical values between 1 and 50
}
\keyword{datasets}
