% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_first_function.r
\name{my_first_function}
\alias{my_first_function}
\title{My First Open Science Function}
\usage{
my_first_function(open_science)
}
\arguments{
\item{open_science}{Do you want to practice Open Science? Default is TRUE.}
}
\description{
This function allows you to practice with Open Science
}
\examples{
my_first_function()
}
