% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_gapminder.R
\docType{data}
\name{my_gapminder}
\alias{my_gapminder}
\title{Edgard Anderson's Iris Data.}
\format{
a data frame with 1704 observations and 6 variables:
\describe{
\item{country}{factor with 142 levels, each one being a country}
\item{continent}{factor with 5 levels}
\item{year}{ranges from 1952 to 2007 in increments of 5 years}
\item{lifeExp}{life expectancy at birth, in years}
\item{pop}{population}
\item{gdpPercap}{Gross Domestic Product per capita (US$, inflation-adjusted)}
}
}
\source{
\url{http://www.gapminder.org/data/}. Gapminder is an independent Sweedish
foundation which claims to "fight devastating misconceptions about global
development".
}
\usage{
my_gapminder
}
\description{
The famous \code{\link[datasets]{iris}} dataset, an excerpt of the Gapminder
data on life expectancy, GDP per capita, and population by country. To be
used in our package.
}
\keyword{datasets}
