% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{K Nearest Neighbours}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{input data frame}

\item{cl}{true class value of your training data}

\item{k_nn}{integer representing the number of neighbors}

\item{k_cv}{integer representing the number of folds}
}
\value{
list containing a vector of the predicted class for all observations,
and a numberic with the cross-validation and misclassification error
}
\description{
Predicts output class species using covariates bill_length_mm, bill_depth_mm,
flipper_length_mm, and body_mass_g.
}
\examples{
my_penguins <- na.omit(my_penguins)
my_knn_cv(my_penguins[,3:6], my_penguins$species, 1, 2)

}
\keyword{prediction}
