% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear Model}
\usage{
my_lm(data, formula)
}
\arguments{
\item{formula}{a formula class object}

\item{data:}{input data frame}
}
\value{
table with rows for each coefficient and columns for the estimate,
standard error, t value, and Pr(>|t|)
}
\description{
This function fits a linear model in R
}
\examples{
# my_lm(formula, x)
# my_lm(formula, y)
x <- 1 : 50
y <- rnorm(50, mean = 25, sd = 10)
z <- rnorm(50, mean = 25, sd = 5)
my_data <- data.frame("X" = x, "Y" = y, "Z" = z)

}
\keyword{inference}
