% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.r
\name{my_t.test}
\alias{my_t.test}
\title{T-test}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{a numeric vector of data}

\item{alternative}{a character string specifying the alternative hypothesis.}

\item{mu}{a number indicating the null hypothesis value of the mean.}
}
\value{
list containing the test statistic, degrees of freedom, value of the
parameter alternative, and the p-value
}
\description{
This function performs a one sample t-test in R.
}
\examples{
x <- 1:10
my_t.test(x, "less", 5)

}
\keyword{inference}
