% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t_test.R
\name{my_t_test}
\alias{my_t_test}
\title{Student's t-test}
\usage{
my_t_test(x, alternative, mu)
}
\arguments{
\item{x}{a numeric vector of data.}

\item{alternative}{a character string specifying the alternative hypothesis.
This only accepts "two.sided", "less", or "greater".}

\item{mu}{a number indicating the null hypothesis value of the mean.}
}
\value{
A list with elements: \code{test_stat}, the numeric test statistic,
\code{df}, a numeric containing the degrees of freedom, \code{alternative}
character containing the value of the parameter alternative, \code{p_val},
a numeric p-value.


test_stat: the numeric test statistic.
df: the degrees of freedom.
alternative: the value of the parameter alternative.
p_val: the numeric p-value.
}
\description{
This function performs a one sample student's t-test on a vector of data.
}
\examples{
f_to_c(32)
f_to_c(212)

}
\keyword{t-test}
