\name{myfirstpackage-package}
\alias{myfirstpackage-package}
\alias{myfirstpackage}
\docType{package}
\title{
\packageTitle{myfirstpackage}
}
\description{
\packageDescription{myfirstpackage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{myfirstpackage}
\packageIndices{myfirstpackage}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{myfirstpackage}

Maintainer: \packageMaintainer{myfirstpackage}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
