% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_one.R
\name{myfunc1}
\alias{myfunc1}
\title{Simples function to sum up two numbers}
\usage{
myfunc1(num1 = numeric(), num2 = numeric())
}
\arguments{
\item{num1}{a numeric value}

\item{num2}{another numeric value}
}
\value{
the sum of two numeric values
}
\description{
Simples function to sum up two numbers
}
\examples{
myfunc1(1+2)
}
