% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myheatmap.R
\name{myheatmap}
\alias{myheatmap}
\title{Plots a heatmap of the N variable genes you find from using the earlier function.}
\usage{
myheatmap(sco)
}
\arguments{
\item{sco}{separated text file with row names and header}
}
\description{
Plots a heatmap of the N variable genes you find from using the earlier function.
}
