% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{myPower}
\alias{myPower}
\title{Power a number}
\usage{
myPower(x, y)
}
\arguments{
\item{x}{A numeric value to be the base}

\item{y}{A numeric value to be the exponent}
}
\value{
the power of the base
}
\description{
Takes a numeric value as the base and another numeric value as the exponent
}
\examples{
power(1,0)
power(2,2)
}
