% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation.R
\name{mySum}
\alias{mySum}
\title{Sum of two numeric values}
\usage{
mySum(x, y)
}
\arguments{
\item{x}{A numeric value}

\item{y}{A numeric value}
}
\value{
the sum of these two numeric values
}
\description{
Sum of two numeric values
}
\examples{
sum(1,2)
sum(3,5)
}
