% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_check.R
\name{n_check}
\alias{n_check}
\title{Creating a function for rejecting negative input values}
\usage{
n_check(r)
}
\arguments{
\item{r, }{Number input}
}
\description{
Creating a function for rejecting negative input values
}
\examples{
n_check(-2)
}
