% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{ndvi}
\alias{ndvi}
\title{Calculate NDVI}
\usage{
ndvi(red, nir)
}
\arguments{
\item{red}{value or vector with reflectance in the Red wavelenght}

\item{nir}{value or vector with reflectance in the NIR wavelenght}
}
\value{
A matrix of the infile
}
\description{
This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kepted.
}
