% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{order.mat}
\alias{order.mat}
\title{Order Statistic for all 2x2 Tables}
\usage{
order.mat(N.T, N.C, Delta0, method)
}
\arguments{
\item{N.T}{sample size in Treatment group}

\item{N.C}{sample size in Control group}

\item{Delta0}{non-inferiority margin}

\item{method}{method for ordering criterion}
}
\value{
(N.T+1)x(N.C+1) array where the (i,j) element is the order statistic for x.T=i and x.C=j
}
\description{
This function calculate order statistic for all 2x2 tables given N.T, N.C and an ordering criterion
}
