% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitbullsFunc.R
\name{pitbullsFunc}
\alias{pitbullsFunc}
\title{An Ignorance test function}
\usage{
pitbullsFunc(love = TRUE)
}
\arguments{
\item{Do}{you love pitbulls? Defaults to TRUE}
}
\description{
This function allows you to determine of you are ignorant by answering one question.
Get a nice message if not and if you are, an explanation why.
}
\examples{
fibonacci( 5 )

}
\keyword{pitbulls}
