% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coati_functions.R
\name{plot_coati}
\alias{plot_coati}
\title{Plot size by age for coatis}
\usage{
plot_coati(x, y)
}
\arguments{
\item{x}{coati age}

\item{y}{coati length}
}
\description{
This function plots coati age on the x axis and length on the y axis
}
\examples{
coatis <- data.frame(name= c("coati1", "coati2", "coati3"),age = c(2,3,4),length = rnorm(3,50,10))
plot_age(coatis$age, coatis$length)
}
