% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_confusion_matrix}
\alias{plot_confusion_matrix}
\title{Plot Confusion Matrix}
\usage{
plot_confusion_matrix(test_data, predicted, labels)
}
\arguments{
\item{test_data}{The testing data}

\item{predicted}{The predicted values}

\item{labels}{The true labels}
}
\value{
A ggplot object representing the confusion matrix plot
}
\description{
This function generates a plot of the confusion matrix based on the predicted
values and true labels.
}
