% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Plot simple histogram}
\usage{
plot_histogram(x)
}
\arguments{
\item{x}{numeric; Numeric vector to plot}
}
\description{
This function shall plot a histgram using base R.
}
\examples{
plot_histogram(rnorm(100))
}
