% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_metrics}
\alias{plot_metrics}
\title{Plot Metrics}
\usage{
plot_metrics(trainer, metrics)
}
\arguments{
\item{trainer}{The trainer object containing the training and test data}

\item{metrics}{The metrics calculated during the learning curve analysis}
}
\value{
If the status is "classification":
\item{acc}{Accuracy plot}
\item{prec}{Precision plot}
\item{rec}{Recall plot}
\item{f1}{F1-score plot}

If the status is "regression":
\item{metrics_df}{Data frame of all metrics}
\item{p_mse}{MSE plot}
\item{p_rmse}{RMSE plot}
\item{p_mae}{MAE plot}
\item{p_rsquared}{R-squared plot}
}
\description{
This function generates plots to visualize the metrics calculated during the learning curve analysis.
The plots differ based on the status of the trainer object (classification or regression).
For classification models, the plots include accuracy, precision, recall, and F1-score.
For regression models, the plots include mean squared error (MSE), root mean squared error (RMSE),
mean absolute error (MAE), and coefficient of determination (R-squared).
}
