% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_probs}
\alias{plot_probs}
\title{Probability Distribution Plot}
\usage{
plot_probs(trainer, probs)
}
\arguments{
\item{trainer}{The trainer object containing the true labels}

\item{probs}{The predicted probabilities obtained from the trained model}
}
\value{
A histogram plot visualizing the distribution of predicted probabilities
}
\description{
This function generates a histogram plot to visualize the distribution of predicted probabilities for each class in a classification problem.
The plot displays the predicted probabilities on the x-axis and the count of data points on the y-axis.
The histogram bins represent the predicted probability ranges, and each class is filled with a different color.
}
