% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfib.R
\name{plotfib}
\alias{plotfib}
\title{A Fibonacci plot function}
\usage{
plotfib(n)
}
\arguments{
\item{n, }{Number of Values returned}
}
\description{
This function allows you to plot N numbers of the
Fibonacci Sequence
}
\examples{
plotfib( 20 )


}
\keyword{plotfib}
