% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare time series data for usage}
\usage{
prepare_data(flows, stocks, inventory, target)
}
\arguments{
\item{flows}{A list of time series variable (flow variables).}

\item{stocks}{A list of time series variable (stock variables).}

\item{inventory}{Inventory of the corresponding data.}

\item{target}{Defines the target variable.}
}
\value{
A dataframe
}
\description{
\code{prepare_data} transforms the time series into a usable format for time series analysis.
}
\details{
The function conducts a normalization by demeaning and dividing by the standard deviation of the individual time series.
}
\examples{
Xmat <- prepare_data(flows = data$flows, stocks = data$stocks, inventory, target = target)
}
