% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coati_functions.R
\name{print_names}
\alias{print_names}
\title{Print coati name}
\usage{
print_names(x)
}
\arguments{
\item{x}{the dataframe containing the coati information. Column must be named "name"}
}
\description{
This function prints coati names
}
\examples{
coatis <- data.frame(name= c("coati1", "coati2", "coati3"),age = c(2,3,4),length = rnorm(3,50,10))
print(coatis)
}
