% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{print_star_wars_greeting}
\alias{print_star_wars_greeting}
\title{Print a Star Wars Theme message}
\usage{
print_star_wars_greeting(who = "Obi Wan Kenobi")
}
\arguments{
\item{who}{should be a character from Star Wars}
}
\description{
Print a Star Wars Theme message
}
