% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation.R
\name{product}
\alias{product}
\title{Product of two numeric values}
\usage{
product(x, y)
}
\arguments{
\item{x}{A numeric value}

\item{y}{A numeric value}
}
\value{
the product of these two numeric values
}
\description{
Product of two numeric values
}
\examples{
product(1,2)
product(2,3)
}
