% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv.R
\name{pv}
\alias{pv}
\title{Present Value}
\usage{
pv(FV, r, n = 5)
}
\arguments{
\item{FV}{a number representing some future value}

\item{r}{rate of return}

\item{n}{number of time periods}
}
\description{
Compute the present value based on a specified future value,
interest rate, and number of time periods.
}
\examples{
pv(FV = 1000, .09, n = 5)

}
