% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation.R
\name{quorient}
\alias{quorient}
\title{Quotient of two numeric values}
\usage{
quorient(x, y)
}
\arguments{
\item{x}{A numeric value}

\item{y}{A numeric value}
}
\value{
the quotient of these two numeric values
}
\description{
Quotient of two numeric values
}
\examples{
quorient(1,2)
quotient(2,3)
quotient(3,3)
}
