% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readfacs.R
\name{read.facs}
\alias{read.facs}
\title{Reads in file and converts it to the scell type}
\usage{
read.facs(sco, fname)
}
\arguments{
\item{sco}{Scell object}

\item{fname}{"" separated text file with row names and header}
}
\description{
Reads in file and converts it to the scell type
}
