% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readmyfile.R
\name{read.myfile}
\alias{read.myfile}
\title{Reads in file and converts it to the scell type}
\usage{
read.myfile(sco)
}
\arguments{
\item{sco}{"" separated text file with row names and header}
}
\description{
Reads in file and converts it to the scell type
}
