% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFACS.R
\name{readFACS}
\alias{readFACS}
\title{Read FACS data and put it into a scell object}
\usage{
readFACS(sco, FileIn)
}
\arguments{
\item{sco}{the scell object}

\item{FileIn}{the facs file}
}
\description{
Read FACS data and put it into a scell object
}
