% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadInS4.R
\name{readInData}
\alias{readInData}
\title{Read in file and transfer it to a S4 object}
\usage{
readInData(inp)
}
\arguments{
\item{inp}{Input file}
}
\description{
Read in file and transfer it to a S4 object
}
