% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readindexfile.R
\name{readindexfile}
\alias{readindexfile}
\title{Read delimeited text file into sclass-object}
\usage{
readindexfile(sco, myfilename)
}
\arguments{
\item{sco}{s4 class object}

\item{myfilename}{input filename}
}
\description{
Read delimeited text file into sclass-object
}
