% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewSCellExpt.R
\name{ReadNewSCell}
\alias{ReadNewSCell}
\title{Make a new scell class from a file}
\usage{
ReadNewSCell(FileIn)
}
\arguments{
\item{FileIn}{An scell S4 object}
}
\description{
Make a new scell class from a file
}
