% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.R
\name{report_p}
\alias{report_p}
\title{Report p values}
\usage{
report_p(p, digits = 3)
}
\arguments{
\item{x}{numeric value of p}

\item{y}{numeric value of sig digits}
}
\value{
reported text
}
\description{
This function takes a p-value and to a specified number of digits reports as text.
}
\examples{
report_p(0.001, 3)

}
