% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\title{Rescale}
\usage{
rescale(x, digits = 2, na.rm = TRUE)
}
\arguments{
\item{x}{a vector of numeric values}

\item{digits}{number of digits to round the output to}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\description{
Normalizes a numeric vector by standardizing the values between 0-1.
}
\examples{
x <- runif(25, min = 5, max = 20)
rescale(x)

}
