% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{restricted.ml}
\alias{restricted.ml}
\title{Restricted MLE of Variance}
\usage{
restricted.ml(x.T, x.C, N.T, N.C, Delta)
}
\arguments{
\item{x.T}{observed number of responders in Treatment group}

\item{x.C}{observed number of responders in Control group}

\item{N.T}{sample size in Treatment group}

\item{N.C}{sample size in Control group}

\item{Delta}{constraint for P.C-P.T}
}
\value{
a vector (MR.T, MR.C) where MR.T and MR.C are the restricted MLE for the Treatment and Control group respectively
}
\description{
This function calculate the maximum likelihood estimator (MLE) of the variance
restricted to P.T-P.C=-Delta
}
