% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellscell.R
\name{SingleCellData}
\alias{SingleCellData}
\title{Reads FACS data and put it into a scell object}
\usage{
SingleCellData(sco, InputFile)
}
\arguments{
\item{InputFile}{Index sort data input file containing expression data}
}
\description{
Reads FACS data and put it into a scell object
}
