% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_of_sq.R
\name{sq_a}
\alias{sq_a}
\title{Square Area}
\usage{
sq_a(l, w)
}
\arguments{
\item{l}{the length of rectangular}

\item{w}{the width of rectangular}
}
\description{
Calculates area of a rectangular
}
\author{
M Usman Mirza
}
