% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_world.R
\name{sqroot}
\alias{sqroot}
\title{Square Root}
\usage{
sqroot(x)
}
\arguments{
\item{x}{the numeric value}
}
\value{
the square root of x
}
\description{
Calculate the square root of a numeric value
}
\examples{
sqroot(22);
}
