% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE.R
\name{std.error}
\alias{std.error}
\title{Calculate standard error of mean}
\usage{
std.error(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a numeric value for standard error
}
\description{
a simple function to calculate standard error of mean
}
\examples{
std.error(rnorm(100))

}
