% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substringIndex.R
\name{substring_index}
\alias{substring_index}
\title{Substring Index}
\usage{
substring_index(string, delimiter, number)
}
\arguments{
\item{number}{}
}
\value{

}
\description{
This function mimics MySQL's substring_index() function.
}
\examples{
mtcars \%>\% rownames_to_column() \%>\% as_tibble() \%>\% mutate(ck = substring_index(rowname, ' ', 1))
}
