% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{subtract}
\alias{subtract}
\alias{minus}
\title{A function for subtracting}
\usage{
subtract(x, y)

minus(x, y)
}
\arguments{
\item{x}{the first number}

\item{y}{the second number}
}
\value{
x-y (numeric)
}
\description{
A decent function

A work of pure genius.
}
\examples{
subtract(5, 10)
## Can also use negative numbers
subtract(-5, 10)
minus(20,20)
minus(9,8)
}
