% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphs.R
\name{ViolinPlotFACSfromCluster}
\alias{ViolinPlotFACSfromCluster}
\title{Make a violin plot of the facs data from particular cluster defined in the clusters slot}
\usage{
ViolinPlotFACSfromCluster(sco, clusID)
}
\description{
Make a violin plot of the facs data from particular cluster defined in the clusters slot
}
