% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_of_sphere.R
\name{volume_of_sphere}
\alias{volume_of_sphere}
\title{Sphere volume}
\usage{
volume_of_sphere(radius)
}
\arguments{
\item{radius}{the radius of the sphere}
}
\description{
Function which calculates volume of sphere
}
\author{
Eveline van der Linden
}
