% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_parameter.R
\name{weibull_par}
\alias{weibull_par}
\title{Estimating weibull parameter}
\usage{
weibull_par(dataset)
}
\arguments{
\item{dataset}{the time series data}
}
\description{
Function that estimates two parameters of Weibull distribution
}
\author{
Yu Jiang
}
