% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic.R
\name{z_score}
\alias{z_score}
\title{Z-score function}
\usage{
z_score(x, y, z)
}
\arguments{
\item{x}{A raw score}

\item{y}{The mean value}

\item{z}{The standard deviation value}
}
\value{
the difference of raw score and mean value divide the standard deviation value
}
\description{
A z-score describes the position of a raw score in terms of its distance from the mean
when measured in standard deviation units
}
\examples{
z_score(190,150,25)
}
