% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\alias{add_quantiles}
\title{Adding Numbers and Square}
\usage{
add(x, y)

add_quantiles(x, y)
}
\arguments{
\item{x}{numeric; The first parameter in my add and square function}

\item{y}{numeric; The second parameter}
}
\description{
Adding Numbers and Square
}
