% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mat.R
\name{add}
\alias{add}
\title{add Math}
\usage{
add(a, b)
}
\arguments{
\item{a}{math parameter 1}

\item{b}{math parameter 2}
}
\value{
result
}
\description{
This function returns a result adds two parameters
}
