% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{area_of_square}
\alias{area_of_square}
\title{Calculate area of square}
\usage{
area_of_square(length, width)
}
\arguments{
\item{length}{The length of the square}

\item{width}{The width of the square}
}
\description{
Calculate area of square
}
