% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tSNE dim3.R, R/tSNE.R
\name{CalcTSNE}
\alias{CalcTSNE}
\title{Calculate a tSNE over 3 dims using the variable genes only}
\usage{
CalcTSNE(sco)

CalcTSNE(sco)
}
\arguments{
\item{sco}{An scell S4 object}

\item{sco}{An scell S4 object}
}
\description{
Calculate a tSNE over 3 dims using the variable genes only

Calculate a tSNE over 3 dims using the variable genes only
}
