% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map with plotted accidents.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A character that represents the state number.}

\item{year}{A character.}
}
\value{
a map with accidents plotted in it. It will also returns a message that there are no accidents to plot
if there no records of accident for the given state.
}
\description{
This function will take \code{state.num} and \code{year} as input and returns a map with accident plotted in it.
It will throw an error if the \code{state.num} is an invalid'.
}
\note{
This function will throws an error if \code{state.num} is invalid.
}
\examples{
\dontrun{
fars_map_state("1", "2013")
	fars_map_state(state.num = "1", year = "2013")
 }

}
