% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Illustrate data with a map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An identifier for the state}

\item{year}{A year of the data which is going to be illustrated}
}
\value{
This function plots the data
}
\description{
This is a function that illustrates data with the given state.num and year by a map.
If data does not contain the given state.num, the execution of the function is stopped
and "invalid STATE number" message is given. If the given state.num exists in the data
but there is no observations, the function gives a message "no accidents
to plot".
}
\examples{
fars_map_state(1, 2013)

}
