% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plots fatal accidents in a State for a particular year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Integer denoting the state}

\item{year}{Integer representing the year of interest}
}
\value{
This function renders and returns a map object
}
\description{
Function that renders on a map the location of fatal accidents that occurred in a specific State in a particular year
}
\details{
Function aborts if the state number state.num is invalid
}
\note{
Makes use of fars_read() and make_filename() functions
}
\examples{
\dontrun{
fars_map_state(20, 2013)
}

}
