% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read a FARS file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character.}
}
\value{
This function returns a dataframe table.
}
\description{
This function reads a file, to be precise a  file from Fatality Analysis Reporting System.
We can pass a file name to read (using the \code{filename} argument).
}
\note{
This function will throws an error if the filename doesn't exist.
}
\examples{
\dontrun{
	fars_read("fars_data.csv")
	fars_read(filename = "fars_data.csv")
	fars_read("data/accident_2013.csv")
}


}
