% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read data from a file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A name of the file that contains the data to be read}
}
\value{
This function returns the data as a tibble object
}
\description{
This function reads a csv-formatted data from a given file.
If the given file does not exist, the execution is stopped and an error message is given.
}
\examples{
fars_read("accident_2013.csv.bz2")
}
