% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{fars_read}
\alias{fars_read}
\title{Read data from file into a tibble}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Name of the csv file}
}
\value{
This function returns a tibble.
}
\description{
Function that reads in data from a csv file and returns a tibble
}
\details{
Function aborts and prints error message if the file filename cannot be found
}
\examples{
fars_read("input_file.csv")

}
