% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Data for the given years.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years.}
}
\value{
A list of dataframes for each year in the vector of years if that year exist,
or it will return a NULL for invalid year.
}
\description{
This function takes a vector of years, and for each year in the vector it uses \code{make_filename(year)} to make a file name pattern
for the given year and \code{fars_read(filename)} to read the file.
}
\note{
This function will throw a warning if the entered year is invalid.
}
\examples{
\dontrun{
 fars_read_years(c("2013", "2014"))
 yearvector <- c("2013", "2014")
 fars_read_years(yearvector)
}

}
