% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read data from files referring to certain years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years}
}
\value{
This function returns the data as a tibble object containing columns MONTH and year
}
\description{
This function reads csv-formatted data from files that refer to given years
ie. they contain data from the given years, If one of the files can not be readed, the function
throws a warning message and returns NULL.
}
\examples{
fars_read_years(c(2019, 2020))
}
