% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize the data by years.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years.}
}
\value{
summarized data that spread across multiple columns.
}
\description{
This function takes a vector of \code{years} as input. Then reads the list generated by \code{fars_read_years(years)} by calling it.
Then summarize the data and spreads the data in multiple column.
}
\examples{
\dontrun{
fars_summarize_years(c("2013", "2014"))
fars_summarize_years(yearvector)
}

}
