% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Presents number of fatal injuries from FARS by month for  multiple years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Vector denoting the years of interest, i.e.,when the yearly FARS data is required}
}
\value{
This function returns a table showing the number of fatal injuries by month, for the years of interest
}
\description{
Function that reads in yearly FARS data, and tabulates the number of fatal injuries by month, for the years of interest
}
\note{
Makes use of fars_read_years() functions
}
\examples{
fars_summarize_years(c("2020", "2019", "2015"))

}
