% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibonacci.R
\name{fibonacci}
\alias{fibonacci}
\title{A fibonacci generator function}
\usage{
fibonacci(n)
}
\arguments{
\item{n, }{Number of Values returned}
}
\description{
This function allows you to generate N numbers of the
Fibonacci Sequnce
}
\examples{
fibonacci( 5 )
[1] 0 1 1 2 3
}
\keyword{fibonacci}
