% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_world.R
\name{hello_world}
\alias{hello_world}
\title{Hello World}
\usage{
hello_world()
}
\value{
"Hello world! sentence"
}
\description{
Print Hello world with any other use
}
\examples{
p <- hello_world();
print(p);
}
