% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make a file name for the given year.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A character.}
}
\value{
This function will return the file name in desired pattern for the given \code{year}.
}
\description{
This function creates a file name in desired pattern by taking a \code{year} as input.
}
\examples{
\dontrun{
 	make_filename("2013")
	make_filename(year = "2014")
}

}
