% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make a filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A certain year to which the filename refers}
}
\value{
This function returns a character vector of a form "accident_YEAR.csv.bz2"
}
\description{
This function makes a name for a file refering to a given year.
That year is given as a parameter to he function.
}
\examples{
make_filename(2020)
}
