% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Generate string with defined format for denoting a filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Year to tag the filename with}
}
\value{
This function returns a string
}
\description{
Function that creates a string accident_XXXX.csv.bz2 where XXXX represents the year
}
\examples{
\dontrun{
make_filename("2020")
}

}
