% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_gapminder.R
\docType{data}
\name{my_gapminder}
\alias{my_gapminder}
\title{Jennifer Bryan's Gapminder Data}
\format{a data frame with 1704 observations and 6 variables:
\describe{
\item{country}{names of countries}
\item{continent}{names of continent}
\item{year}{year data was taken from}
\item{lifeExp}{average life expectancies}
\item{pop}{total population}
\item{gdpPercap}{GDP per capita}
}}
\source{
http://www.gapminder.org/data/
}
\usage{
my_gapminder
}
\description{
Excerpt of the Gapminder data on life expectancy, GDP per capita, and
  population by country.
}
\keyword{datasets}
