% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_gapminder.R
\docType{data}
\name{my_gapminder}
\alias{my_gapminder}
\title{Gapminder Data}
\format{
a data frame with 1704 observations and 6 variables:
\describe{
\item{country}{factor with 142 levels}
\item{continent}{factor with 5 levels}
\item{year}{ranges from 1952 to 2007 in increments of 5 years}
\item{lifeExp}{life expectancy at birth, in years}
\item{pop}{population}
\item{gdpPercap}{GDP per capita (US$, inflation-adjusted)}
}
}
\source{
\url{http://www.gapminder.org/data/}
}
\usage{
my_gapminder
}
\description{
Excerpt of the \code{Gapminder} data on life expectancy, GDP per capita, and
population by country.
}
\keyword{datasets}
