% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{K-Nearest Neighbors function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{An input data frame.}

\item{cl}{The true class value of the training data.}

\item{k_nn}{An integer representing the number of neighbors to use.}

\item{k_cv}{An integer representing the number of folds to use.}
}
\value{
A list with objects \code{class} (a vector of the predicted class
for all observations) and \code{cv_err} (a numeric with the
cross-validation misclassification error).
}
\description{
This function performs predicts an output class using covariates
given in training data.
}
\examples{
train_dat <- subset(na.omit(my_penguins), select = c(bill_length_mm,
  bill_depth_mm, flipper_length_mm, body_mass_g))
correct <- data.frame("species" = na.omit(my_penguins)$species)
my_knn_cv(train_dat, correct, 1, 5)

}
\keyword{prediction}
