% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{K-nearest neighbors cross-validation}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Data set to train model on.}

\item{cl}{Vector of values for the true class.}

\item{k_nn}{Numeric representing the number of neighbors to base model on.}

\item{k_cv}{The number of folds to be used in the model.}
}
\value{
List containg vector of predicted classes, and numerics of the
  cross-validation and training errors.
}
\description{
This function creates a knn model as well as calculating cross-validation
  and training errors
}
\examples{
my_knn_cv(cbind(my_gapminder$gdpPercap, my_gapminder$lifeExp),
          my_gapminder$continent, 10, 5)
my_knn_cv(my_iris[, -5], my_iris$Species, 2, 5)

}
\keyword{prediction}
