% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{K nearest neighbors}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{input dataframe}

\item{cl}{true class value of your training data}

\item{k_nn}{integer value representing the number of neighbors}

\item{k_cv}{integer value representing the number of folds}
}
\value{
A list with the predicted class values and the cross validation misclassification error
}
\description{
Uses k nearest neighbors as a method of statistical prediction and records the error
}
\examples{
my_penguins <- my_penguins \%>\% na.omit
my_knn_cv(my_penguins[3:6], my_penguins$species, 1, 5)

}
\keyword{k}
\keyword{nearest}
\keyword{neighbors}
