% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-Nearest Neighbors Cross-Validation}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{An input data frame with all the values of the predictor
variables.}

\item{cl}{The (actual) response values corresponding to each of the values
from \code{train}.}

\item{k_nn}{A numeric integer representing the number of neighbors
considered.}

\item{k_cv}{A numeric integer representing the number of folds for which to
divide the observations in \code{train}.}
}
\value{
A list with elements \code{class} and \code{cv_err}, where
  \code{class} is a vector of the predicted classes for all
  observations, and \code{cv_err} is a numeric with the cross-validation
  misclassification error.
}
\description{
This function predicts the output class from a given set of covariates and
actual class values using the k-nearest neighbors approach of identifying
patterns in data points near each other.
}
\examples{
my_knn_cv(my_iris[, 1:4], my_iris[, 5], 1, 5)
my_knn_cv(my_iris[, 1:4], my_iris[, 5], 5, 5)

}
\keyword{prediction}
