% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{K-Nearest Neighbors Cross-Validation.}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{cl}{true class value of training data.}

\item{k_nn}{integer representing the number of neighbors.}

\item{k_cv}{integer representing the number of folds.}
}
\value{
A list with objects \code{class}, a vector of the predicted class
y_hat when all the data is used for both training and testing, and
\code{cv_err} a numeric with the average cross-validation misclassification
errorr.
}
\description{
K-Nearest Neighbors Cross-Validation.
}
\examples{
my_knn_cv(train = my_gapminder[, c(4, 6)], cl = my_gapminder$continent, k_nn = 5, k_cv = 5)

}
\keyword{prediction}
