% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear Model function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{a formula class object, similar to lm().}

\item{data}{An input data frame.}
}
\value{
A dataframe with rows for each coefficient and columns for
the \code{Estimate}, \code{Std. Error}, \code{t value}, and \code{Pr(>|t|)}
with row and column names.
}
\description{
This function fits a linear model.
}
\examples{
my_lm(y ~ x, data.frame(x = rnorm(100, 0, 1), y = rnorm(100, 0, 1)))

}
\keyword{inference,}
\keyword{prediction}
