% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear Model}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{a formula class object}

\item{data}{input dataframe}
}
\value{
A table with rows for each coefficient and columns for the estimate, standard error, test statistic and p-value
}
\description{
This function creates a linear model and tests if there is a correlation between the model and the data.
}
\examples{
my_lm(bill_length_mm~bill_depth_mm,my_penguins)


}
\keyword{inference}
