% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear Model}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{An object of class `formula`, which represents the
overall relationship between our predictor variable(s) from \code{data} and
our response variable from \code{data}.}

\item{data}{An input data frame from which \code{formula} extracts from.}
}
\value{
A `table` with rows for the coefficients of each predictor variable
  (including the intercept) and columns for their estimated value, standard
  error, t-value, and p-value (`Estimate`, `Std. Error`, `t value`, and
  `Pr(>|t|)` respectively).
}
\description{
This function performs multiple regression using a given formula and dataset.
}
\examples{
my_lm(Sepal.Length ~ Sepal.Width + Petal.Length + Petal.Width, my_iris)
my_lm(lifeExp ~ gdpPercap + continent, my_gapminder)

}
\keyword{inference}
\keyword{prediction}
