% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Fitting Linear Models}
\usage{
my_lm(formula_lm, data_lm)
}
\arguments{
\item{formula}{a \code{formula} class object: a symbolic description of the
model to be fitted.}

\item{data}{input data frame.}
}
\value{
A list with elements: \code{test_stat}, the numeric test statistic;
\code{df}, a numeric containing the degrees of freedom; \code{alternative}
character containing the value of the parameter alternative; \code{p_val},
a numeric p-value.
}
\description{
This function is used to fit linear models. It can be used to carry out
regressions.
}
\examples{
t_test(c(1, 5 ,4 , 5, 2, 6, 4), greater, 5)
t_test(1:20, two.sided, 18)

}
\keyword{t-test}
