% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Fitting Linear Models}
\usage{
my_lm(formula_lm, data_lm)
}
\arguments{
\item{formula_lm}{formula}

\item{data_lm}{dtatset}
}
\description{
This function is used to fit linear models. It can be used to carry out
regressions.
}
