% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest cross-validation}
\usage{
my_rf_cv(train, k)
}
\arguments{
\item{train}{Data set to train model on}

\item{k}{Numeric input representing the number of folds to be use as
parameters}
}
\value{
List of numeric representing the average MSE error across the folds.
}
\description{
This function calculates MSE erros of a random forest prediction of given
  data.
}
\examples{
my_rf_cv(my_gapminder, 2)
my_rf_cv(my_iris, 5)

}
\keyword{prediction}
