% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{an integer representing the number of folds}
}
\value{
Numeric with the cross validation error
}
\description{
This function uses Random Forest as a method of statistical prediction and records the error
}
\examples{
my_rf_cv(5)

}
\keyword{Forest}
\keyword{Random}
