% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross-Validation}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Numeric indicating the number of folds with which to separate our
dataset.}
}
\value{
Numeric representing the average Mean Squared Error across all
  \code{k} folds (the cross-validation error).
}
\description{
This function uses the \code{my_gapminder} dataset and predicts the Life
Expectancy at birth (in years) from GDP per capita, using Random Forest
Cross-Validation.
}
\examples{
my_rf_cv(4)
my_rf_cv(10)

}
\keyword{prediction}
