% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross-Validation.}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{number of folds.}
}
\value{
Return the average Mean Squared Error across all k folds.
}
\description{
my_rf_cv implements Breiman's random forest algorithm (based on Breiman and
Cutler's original Fortran code) for classification and regression. The
function runs the algoritms for the given number of folds and implements the
algoritm on my_gapminder, predicting lifeExp using covariate gdpPercap.
}
\examples{
my_rf_cv(4)
my_rf_cv(2)

}
\keyword{prediction}
