% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross-Validation}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{something something}
}
\value{

}
\description{
description
}
\examples{
my_rf_cv(4)
my_rf_cv(10)

}
\keyword{}
