% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{One-sample t-test function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{A numeric vector of data.}

\item{alternative}{A character string specifying the alternative hypothesis,
should only be \code{"two.sided"}, \code{"less"}, or \code{"greater"}.}

\item{mu}{A number indicating the null hypothesis value of the mean.}
}
\value{
A list containing four elements: \code{test_stat} (the numeric test,
test statistic), \code{df} (the degrees of freedom), \code{alternative}
(the value of the parameter "alternative"), and \code{p_val}
(the numeric p-value).
}
\description{
This function performs a one-sample t-test.
}
\examples{
my_t.test(x = rnorm(100, 0, 1), alternative = "two.sided", mu = 0)
my_t.test(x = runif(20, 5, 3), alternative = "less", mu = 4)
my_t.test(x = rbinom(70, 6, 0.75), alternative = "greater", mu = 2)

}
\keyword{inference}
