% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t_test.R
\name{my_t.test}
\alias{my_t.test}
\title{t-Test Application}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{Numeric vector representing sample data values.}

\item{alternative}{String that indicates the alternative hypothesis of the
t-test, must be one of "two.sided", "greater" or "less".}

\item{mu}{Numeric indicating the value of the null hypothesis value of the
mean.}
}
\value{
List with elements \code{test_stat}, the value of the t-statistic,
  \code{df}, the degrees of freedom for the t-statistic, \code{alternative},
  the value of the parameter \code{alternative}, and \code{p_val}, the
  numeric p-value that was calculated by the t-test.
}
\description{
This function performs a one sample t-test.
}
\examples{
my_t.test(1:10, "two.sided", 8)
my_t.test(1:10, "less", 3)

}
\keyword{inference}
