% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t_test.R
\name{my_t_test}
\alias{my_t_test}
\title{T test}
\usage{
my_t_test(x, alternative, mu)
}
\arguments{
\item{x}{a numeric vector of data}

\item{alternative}{a character string specifying the alternative hypothesis. Only accepts "two.sided", "less", or "greater".}

\item{mu}{a number indicating the null hypothesis value of the mean.}
}
\value{
A list with the numeric test statistic, the degrees of freedom, the value of the parameter "alternative", and the p value
}
\description{
This function performs a t-test
}
\examples{
my_t_test(rnorm(36), "two.sided", 0)
my_t_test(rpois(40,lambda=8), "less", 7)

}
\keyword{inference}
