% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t_test.R
\name{my_t_test}
\alias{my_t_test}
\title{Fahrenhite converter}
\usage{
my_t_test(x, alternative, mu)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less". You can specify
just the initial letter.}

\item{mu}{}
}
\value{
Numeric indicating temperature \code{temp_F} converted to Celsius.
}
\description{
This function performs a one sample student's t-test on a vector of data.
}
\examples{
f_to_c(32)
f_to_c(212)

}
\keyword{t-test}
