% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t_test.R
\name{my_t_test}
\alias{my_t_test}
\title{Student's t-test}
\usage{
my_t_test(x, alternative, mu)
}
\arguments{
\item{x}{a numeric vector of data.}

\item{alternative}{a character string specifying the alternative hypothesis.
This only accepts "two.sided", "less", or "greater".}

\item{mu}{a number indicating the null hypothesis value of the mean.}
}
\value{
A list with elements: test_stat, the numeric test statistic;
df, a numeric containing the degrees of freedom; alternative
character containing the value of the parameter alternative; p_val,
a numeric p-value.
}
\description{
This function performs a one sample student's t-test on a vector of data.
}
\examples{
my_t_test(c(1, 5 ,4 , 5, 2, 6, 4), "greater", 5)
my_t_test(1:20, "two.sided", 18)

}
\keyword{inference}
