% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFirstPackage-package.R
\docType{package}
\name{myFirstPackage-package}
\alias{myFirstPackage}
\alias{myFirstPackage-package}
\title{myFirstPackage: A Toy Package for Illustration Purposes}
\description{
This is a minimal package with one function and one data frame that we use to illustrate the creation of an R package.
}
\author{
\strong{Maintainer}: Antonio Paez \email{paezha@mcmaster.ca} (\href{https://orcid.org/0000}{ORCID})

}
\keyword{internal}
