\name{MyFirstPackage-package}
\alias{MyFirstPackage-package}
\alias{MyFirstPackage}
\docType{package}
\title{
\packageTitle{MyFirstPackage}
}
\description{
\packageDescription{MyFirstPackage}
}

\author{
\packageAuthor{MyFirstPackage}

Maintainer: \packageMaintainer{MyFirstPackage}
}

\examples{
IC.z.test(datasets, sd(dataset), 0.95)

}
