% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyFirstPackage-package.R
\docType{package}
\name{MyFirstPackage-package}
\alias{MyFirstPackage}
\alias{MyFirstPackage-package}
\title{MyFirstPackage: A Toy Package for Illustration Purposes}
\description{
This is a minimal package with one function and one data frame that we use to Illustrate the creation of an R package.
}
\author{
\strong{Maintainer}: Sara Saffar \email{gharouns@mcmaster.com} (\href{https://orcid.org/0000}{ORCID})

}
\keyword{internal}
