% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{ndvi}
\alias{ndvi}
\title{Calculate NDVI}
\usage{
ndvi(red, nir)
}
\arguments{
\item{red}{value or vector with reflectance in the Red wavelenght}

\item{nir}{value or vector with reflectance in the NIR wavelenght}
}
\value{
NDVI value
}
\description{
This function caculates the normalized difference vegetation index.
}
