% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_world.R
\name{square}
\alias{square}
\title{Square}
\usage{
square(x)
}
\arguments{
\item{x}{the numeric value}
}
\value{
the square of x
}
\description{
Calculate the square of a numeric value
}
\examples{
square(22);
}
