% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substringIndex.R
\name{substring_index}
\alias{substring_index}
\title{Substring Index}
\usage{
substring_index(string, delim, count)
}
\arguments{
\item{number}{}
}
\value{

}
\description{
This function mimics MySQL's substring_index() function.
}
